+++ wayland-server/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+#### Additions
+
 - Add `system` and `dlopen` features, enabling `wayland-backend` features
+- Add `DisplayHandle::set_default_max_buffer_size` and `Client::set_max_buffer_size`
 
 ## 0.31.10 - 2025-07-28